// frontend/magic-web/src/opensource/types/flow.ts
export type PlatformItem = {
	id: string
	name: string
	avatar: string
}
// ...
	export type Detail = {
		id: string
		name: string
		description: string
		creator: string
		created_at: string
		modifier: string
		updated_at: string
		magic_flow: MagicFlow.Flow
		creator_info: PlatformItem
		modifier_info: PlatformItem
	}

	export type RequestArgs = Partial<Pick<Detail, "name" | "description" | "magic_flow" | "id">>

	export type ListItem = Detail
// ...
export enum FlowRouteType {
	Agent = "agent",
	Sub = "sub",
	Tools = "tools",
	VectorKnowledge = "vector-knowledge",
	Mcp = "mcp",
}

export namespace Flow {
	export interface VLMModel {
		id: string
		name: string
		model_id: string
		model_version: string
		model_type: number
		category: string
		icon: string
		description: string
	}

	export interface VLMProvider {
		id: string
		name: string
		models: VLMModel[]
	}

	export enum VLMModelType {
		/** 文生图 */
		TextToImage = 0,
		/** 图生图 */
		ImageToImage = 1,
		/** 图片增强 */
		ImageEnhance = 2,
		/** 大模型 */
		LLM = 3,
		/** 嵌入 */
		Embedding = 4,
	}

	export namespace Mcp {
		export type Detail = {
			created_at: string
			created_uid: null
			creator: string
			creator_info: PlatformItem
			description: string
			enabled: boolean
			id: string
			mcp_server_code: string
			type: string
			modifier: string
			modifier_info: PlatformItem
			name: string
			options: object
			rel_code: string
			rel_version_code: string
			source: number
			updated_at: string
			updated_uid: null
			version: string
			source_version: {
				latest_version_code: string
				latest_version_name: string
			}
			icon: string
			user_operation: OperationTypes
			require_fields?: Array<{ field_name?: string; field_value?: string }>
			[property: string]: any
		}
		export type ListItem = Mcp.Detail
		export type SaveParams = {
			id?: string
			source?: Flow.Mcp.ToolSource
			rel_code?: string
			rel_version_code?: string
			name?: string
			description?: string
			enabled?: boolean
			rel_info?: Record<string, any>
			icon?: string
			type?: string
			external_sse_url?: string
		}
		export type GetListParams = {
			page: number
			pageSize: number
			name: string
		}
		export enum ToolSource {
			/** 工具集 */
			Toolset = 1,
		}
	}

	export enum ApiKeyType {
		None = 0,
		Flow = 1,
		Mcp = 2,
	}

	export type ApiKeyRequestParamsV1 = {
		id?: string
		name?: string
		description?: string
		rel_type?: ApiKeyType
		rel_code?: string
	}
}

/**
 * 向量知识库
 */
export namespace VectorKnowledge {
	export enum SearchType {
		/** 全部 */
		All = 1,
		/** 已启用 */
		Enabled = 2,
		/** 已禁用 */
		Disabled = 3,
	}
}